Restarting the project
I am remaking the project from the ground up as to organize and make it
more akin to a guided tutorial. The original files were more of a
personal experiment, so there was very little organization. Now, I plan
on make a web book about computer graphics, so I'll be remaking
everything from the ground up.